be/requirements.txt
fastapi
uvicorn
pydantic
pytest
python-dotenv
google-generativeai
python-multipart
# MySQL support
SQLAlchemy==2.0.27
PyMySQL==1.1.0
cryptography==41.0.7  # For PyMySQL
alembic==1.13.1  # For database migrations
